fix e2e specs: real join, WebGL stub, correct selectors
Make the e2e specs actually pass against a live Foundry (they had never been
run before):

- Add shared tests/e2e/helpers/game-page.js with the WebGL stub (extracted
  from bootstrap), a correctly-positioned waitForGameReady, and openGamePage
  which performs a real join. The Foundry join is server-side session state
  that does not survive in storageState, so reusing the saved cookie lands on
  /join; each spec context must join via the form.
- Specs and bootstrap now share the stub helper instead of duplicating it.
- Fix actor/item sheet selectors: the sheet/actor/item classes live on the
  window container, not the <form>; the form carries paranoia + actor type
  (and paranoia-item for items).
- Raise the global Playwright timeout to 120s so beforeAll hooks have room to
  reach game.ready (the previous { timeout } arg was in the wrong position and
  the 30s hook cap killed them early).
- Decouple the CI system mount: copy the system into an isolated dir instead
  of bind-mounting the live workspace, which felddy v12 chowns and makes the
  runner's tests/e2e/.auth unwritable.
