+++ .github/workflows/e2e.yml
@@ -46,9 +46,25 @@ jobs:
 
       # The compendium packs (LevelDB) are gitignored and built from the
       # tracked unpacked_packs/ JSON sources. They must exist before Foundry
-      # loads the system, so build them before mounting the system.
+      # loads the system, so build them before staging the system.
       - name: Build compendium packs
         run: npm run pullJSONtoLDB
+
+      # Copy the system into an isolated dir rather than bind-mounting the live
+      # workspace. The felddy entrypoint chowns /data recursively on some Foundry
+      # versions (v12), which would change ownership of the mounted repo and make
+      # tests/e2e/.auth and test-results unwritable by the runner. Copying keeps
+      # the runner's workspace untouched.
+      - name: Stage system for Foundry
+        run: |
+          mkdir -p /tmp/paranoia-system
+          rsync -a --delete \
+            --exclude node_modules \
+            --exclude .git \
+            --exclude test-results \
+            --exclude tests \
+            "$GITHUB_WORKSPACE"/ /tmp/paranoia-system/
+          chmod -R 777 /tmp/paranoia-system
 
       - name: Restore Foundry binary cache
         uses: actions/cache@d4323d4df104b026a6aa633fdb11d772146be0bf # v4.2.2
@@ -61,7 +77,7 @@ jobs:
         # value containing shell metacharacters (quotes, $, backticks) would
         # otherwise break parsing or allow script injection. `docker run -e VAR`
         # (no value) inherits the variable by name from this env block.
-        # Mount the checked-out repo as the paranoia system, mirroring the
+        # Mount the staged system copy as the paranoia system, mirroring the
         # devcontainer (..:/data/Data/systems/paranoia), so createWorld can
         # find the system.
         env:
@@ -77,7 +93,7 @@ jobs:
             --user root \
             -p 30000:30000 \
             -v /tmp/foundryvtt-cache:/data/container_cache \
-            -v "$GITHUB_WORKSPACE":/data/Data/systems/paranoia \
+            -v /tmp/paranoia-system:/data/Data/systems/paranoia \
             -e FOUNDRY_USERNAME \
             -e FOUNDRY_PASSWORD \
             -e FOUNDRY_LICENSE_KEY \
